Added `previewContainer` and scannerOverlay to the viewController.
Added two properties and a method.

Set a UIView object as a container of the preview camera view.
If it is nil, the root view of this ViewController will be used.

Specifies a rectangle of interest for limiting the search area for
visual metadata.

called before initialization. Subclasses can override to set up related
properties.

An overlay UIView to indicate with colored lines where the scanner reads.